Products

+++ app/models/product.py
@@ -6,7 +6,7 @@ class Product(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     product_name = db.Column(db.String(50), nullable=False)
     product_description=db.Column(db.Text, nullable=False)
-    product_price=db.Column(db.Numeric, nullable=False)
+    product_price=db.Column(db.Float, nullable=False)
     product_quantity=db.Column(db.Integer, nullable=False)
     user_id=db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
     user=db.relationship('User', back_populates='products')
